Moves SubscriptionsTabs to the lib
