+++ libs/orchestrator-ui-components/src/lib/components/SubscriptionsTabs/SubscriptionsTabs.tsx
@@ -1,0 +1,47 @@
+import { EuiTab, EuiTabs } from '@elastic/eui';
+import { FC } from 'react';
+
+export enum SubscriptionsTabType {
+    ACTIVE = 'ACTIVE',
+    TERMINATED = 'TERMINATED',
+    TRANSIENT = 'TRANSIENT',
+    ALL = 'ALL',
+}
+
+export type SubscriptionsTab = {
+    id: SubscriptionsTabType;
+    name: string;
+    alwaysOnFilter?: [string, string];
+};
+
+export type SubscriptionsTabsProps = {
+    tabs: SubscriptionsTab[];
+    selectedSubscriptionsTab: SubscriptionsTabType;
+    onChangeSubscriptionsTab: (
+        updatedSubscriptionsTab: SubscriptionsTabType,
+    ) => void;
+};
+
+export const SubscriptionsTabs: FC<SubscriptionsTabsProps> = ({
+    tabs,
+    selectedSubscriptionsTab,
+    onChangeSubscriptionsTab,
+}) => {
+    return (
+        <EuiTabs>
+            {tabs.map(({ id, name }) => (
+                <EuiTab
+                    key={id}
+                    isSelected={id === selectedSubscriptionsTab}
+                    onClick={() => {
+                        if (id !== selectedSubscriptionsTab) {
+                            onChangeSubscriptionsTab(id);
+                        }
+                    }}
+                >
+                    {name}
+                </EuiTab>
+            ))}
+        </EuiTabs>
+    );
+};
