+++ libs/orchestrator-ui-components/src/lib/components/SubscriptionsTabs/index.ts
@@ -1,0 +1,2 @@
+export * from './SubscriptionsTabs';
+export * from './getSubscriptionsTabTypeFromString';
