+++ libs/orchestrator-ui-components/src/lib/components/index.ts
@@ -4,5 +4,6 @@ export * from './OrchestratorPageTemplate';
 export * from './SearchBar';
 export * from './StartPage';
 export * from './Subscription';
+export * from './SubscriptionsTabs';
 export * from './Table';
 export * from './Loading/Loading';
